refactor(lxmf-wire): move the daemon's stamp and ticket helpers into the library
`reticulumd::lxmf_stamps` carried the delivery-stamp, ticket and
peering-key halves of `LXStamper`/`LXMessage` — `generate_stamp`,
`validate_stamp` with tickets, `ticket_stamp`, `COST_TICKET`,
`TICKET_LENGTH`, the peering-key pair and the cancellable generators —
while `lxmf-wire::stamp` exposed only the propagation half and kept
`stamp_workblock`/`stamp_valid`/`stamp_value` private. A library
consumer wanting Python-parity delivery stamps or tickets had to
re-implement them.

The whole set now lives in `lxmf-wire::stamp` (`stamp/mod.rs` for the
shared primitives and propagation stamps, `stamp/delivery.rs` for
delivery stamps, tickets and peering keys), with the `LXMessage` ticket
lifetime constants beside it. The daemon module re-exports the same names,
so its callers are unchanged. Every generator now shares the
`MAX_STAMP_COST` fail-fast the propagation one already had.

Tests move with the code; `ticket_stamp` gains a byte-for-byte vector
from the pinned Python reference and `validate_stamp` a tickets-first
check.
